LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_space-expected.txt

Blocked access to external URL http://www.localhost:8800/resource-timing/resources/TAOResponse.py?tao=space
CONSOLE MESSAGE: line 13: NetworkError:  A network error occurred.
Description

This test validates that for a cross origin resource, the timing allow check algorithm will fail if the Timing-Allow-Origin header value list contains [origin|*]{2,}, separated by space.


PASS window.performance is defined 
PASS window.performance.getEntriesByType is defined 
